Test Suite Debugging P16

+++ lib/ShinyCMS/Controller/Pages.pm
@@ -74,7 +74,10 @@ sub get_section : Chained( 'base' ) : PathPart( '' ) : CaptureArgs( 1 ) {
 		url_name => $section_url_name,
 		hidden   => 0,
 	})->single;
-
+	# KBAKER 20251114: check if url_name exists and if not, trigger alert; delete when not needed
+	unless ($section_url_name eq "NO_SUCH_SECTION")  {
+		die "no section for $section_url_name" unless $c->stash->{ section } ;
+	}
 	# 404 handler
 	$c->detach( 'Root', 'default' ) unless $c->stash->{ section };
 }
